# backend/src/docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Autonomous Content-to-Action Agent — Backend API",
    "version": "0.5.0",
    "description": "Phase 5 documentation surface. Exposes ingestion, pipeline trace, HITL execution gate, contract registry, and AMCE (Modules 11..14) contract shapes. All endpoints are read-only except POST /api/pipeline/run and POST /api/execution/:id/approve."
  },
  "servers": [
    {
      "url": "http://localhost:8000",
      "description": "Local dev server"
    }
  ],
...
  "paths": {
...
    "/api/pipeline/run": {
      "post": {
        "tags": [
          "Pipeline"
        ],
        "summary": "Submit a multi-source pipeline run.",
        "description": "Accepts one or more ingestion sources plus optional constraints, allocates a pipeline_id (PIPE-XXXXXXXX), and (Phase 4+) routes the request to the orchestrator. The current handler returns a synchronous stub envelope; trace/HITL state is observable via the other endpoints once the orchestrator runs.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PipelineRunRequest"
              },
              "examples": {
                "singleUrl": {
                  "summary": "Single URL source",
                  "value": {
                    "sources": [
                      {
                        "source_id": "SRC-001",
                        "source_type": "url",
                        "content": "https://example.com/incident-2026-05-15",
                        "ingested_at": "2026-05-16T14:21:42.001Z"
                      }
                    ],
                    "constraints": {
                      "max_actions": 5
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Run accepted (stub envelope while orchestrator is wired).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PipelineRunResponse"
                }
              }
            }
          },
          "400": {
            "description": "sources array missing or empty.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                },
                "example": {
                  "error": "sources array is required and must not be empty"
                }
              }
            }
          },
          "500": {
            "description": "Unhandled server error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
    "/api/pipeline/{id}": {
      "get": {
        "tags": [
          "Pipeline"
        ],
        "summary": "Full pipeline trace.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^PIPE-[A-Z0-9]{8}$"
            },
            "example": "PIPE-7K2A9F31",
            "description": "Pipeline identifier returned by POST /api/pipeline/run."
          }
        ],
        "responses": {
          "200": {
            "description": "Pipeline trace.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PipelineTrace"
                }
              }
            }
          },
          "404": {
            "description": "Pipeline id not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                },
                "example": {
                  "error": "Pipeline not found"
                }
              }
            }
          }
        }
      }
    },
    "/api/pipeline/{id}/trace": {
      "get": {
        "tags": [
          "Pipeline"
        ],
        "summary": "Antigravity trace export (judge-facing).",
        "description": "Identical shape to GET /api/pipeline/:id; provided as a stable export alias.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^PIPE-[A-Z0-9]{8}$"
            },
            "example": "PIPE-7K2A9F31",
            "description": "Pipeline identifier returned by POST /api/pipeline/run."
          }
        ],
        "responses": {
          "200": {
            "description": "Pipeline trace export.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PipelineTrace"
                }
              }
            }
          },
          "404": {
            "description": "Pipeline trace not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
    "/api/pipeline/{id}/stream": {
      "get": {
        "tags": [
          "Pipeline"
        ],
        "summary": "Server-Sent Events stream of pipeline TraceEvent frames.",
        "description": "Long-lived text/event-stream connection. Each `data:` frame is a JSON-serialized TraceEvent. Emits an `end` event when the trace is finalized. Heartbeat comments keep the connection alive while the pipeline is starting.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^PIPE-[A-Z0-9]{8}$"
            },
            "example": "PIPE-7K2A9F31",
            "description": "Pipeline identifier returned by POST /api/pipeline/run."
          }
        ],
        "responses": {
          "200": {
            "description": "Event stream open. Frames are TraceEvent JSON objects.",
            "content": {
              "text/event-stream": {
                "schema": {
                  "$ref": "#/components/schemas/TraceEvent"
                }
              }
            }
          }
        }
      }
    },
    "/api/pipeline/{id}/analytics": {
      "get": {
        "tags": [
          "Pipeline"
        ],
        "summary": "Compiled analytics for a pipeline.",
        "description": "404 while the multi-agent pool is still synthesizing; 200 once the AMCE-judged PipelineAnalytics record lands in the cache.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^PIPE-[A-Z0-9]{8}$"
            },
            "example": "PIPE-7K2A9F31",
            "description": "Pipeline identifier returned by POST /api/pipeline/run."
          }
        ],
        "responses": {
          "200": {
            "description": "Analytics ready.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PipelineAnalytics"
                }
              }
            }
          },
          "404": {
            "description": "Sub-agents still processing.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ApprovalState": {
        "type": "string",
        "enum": [
          "PENDING",
          "EXECUTING",
          "REJECTED",
          "COMPLETED"
        ],
        "description": "State machine for the HITL approval gate. Only PENDING -> EXECUTING is reachable via POST /api/execution/:id/approve (atomic check-and-set).",
        "example": "PENDING"
      },
...
      "ErrorEnvelope": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "string",
            "example": "Pipeline PIPE-7K2A9F31 is EXECUTING, not PENDING"
          },
          "pipeline_id": {
            "type": "string",
            "pattern": "^PIPE-[A-Z0-9]{8}$",
            "nullable": true
          },
          "current_state": {
            "$ref": "#/components/schemas/ApprovalState"
          }
        }
      },
...
      "MultiSourceIngestionSource": {
        "type": "object",
        "required": [
          "source_id",
          "source_type",
          "content",
          "ingested_at"
        ],
        "properties": {
          "source_id": {
            "type": "string",
            "example": "SRC-001"
          },
          "source_type": {
            "type": "string",
            "enum": [
              "pdf",
              "url",
              "csv",
              "txt",
              "json",
              "realtime_feed"
            ],
            "example": "url"
          },
          "content": {
            "type": "string",
            "example": "https://example.com/incident-report"
          },
          "credibility_tier": {
            "type": "string",
            "enum": [
              "HIGH",
              "MEDIUM",
              "LOW",
              "UNVERIFIED"
            ],
            "example": "HIGH"
          },
          "ingested_at": {
            "type": "string",
            "format": "date-time",
            "example": "2026-05-16T14:21:42.001Z"
          }
        }
      },
...
      "TraceEvent": {
        "type": "object",
        "required": [
          "event_id",
          "pipeline_id",
          "timestamp",
          "event_type",
          "agent",
          "message"
        ],
        "properties": {
          "event_id": {
            "type": "string",
            "format": "uuid"
          },
          "pipeline_id": {
            "type": "string",
            "pattern": "^PIPE-[A-Z0-9]{8}$"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          },
          "event_type": {
            "type": "string",
            "enum": [
              "agent_start",
              "agent_complete",
              "llm_call",
              "contract_gate",
              "action_execute",
              "failure",
              "ingestion_error",
              "recovery",
              "decision",
              "action_start",
              "action_complete",
              "graph_cycle_detected",
              "hitl_pending",
              "hitl_approved",
              "thinking"
            ]
          },
          "agent": {
            "type": "string",
            "example": "StrategicRecommenderAgent"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "type": "object",
            "additionalProperties": true
          },
          "decision": {
            "type": "string"
          },
          "confidence": {
            "type": "number",
            "minimum": 0,
            "maximum": 1
          },
          "provider": {
            "type": "string",
            "example": "gemini-free"
          },
          "latency_ms": {
            "type": "number",
            "minimum": 0
          }
        }
      },
      "ForecastPoint": {
        "type": "object",
        "required": [
          "timestamp",
          "value",
          "is_extrapolation"
        ],
        "properties": {
          "timestamp": {
            "type": "string",
            "format": "date-time"
          },
          "value": {
            "type": "number"
          },
          "is_extrapolation": {
            "type": "boolean",
            "description": "True beyond the current cursor (predicted region)."
          }
        }
      },
      "ContradictionRecord": {
        "type": "object",
        "required": [
          "source_id",
          "timestamp",
          "raw_claim",
          "baseline_context",
          "conflict_rationale"
        ],
        "properties": {
          "source_id": {
            "type": "string",
            "example": "SRC-001"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          },
          "raw_claim": {
            "type": "string"
          },
          "baseline_context": {
            "type": "string"
          },
          "conflict_rationale": {
            "type": "string"
          }
        }
      },
      "PipelineAnalytics": {
        "type": "object",
        "required": [
          "extrapolation_unreliable",
          "forecast_data",
          "contradictions"
        ],
        "description": "AMCE-judged analytics payload returned by GET /api/pipeline/:id/analytics. forecast_data carries 30 historical + 60 predicted points.",
        "properties": {
          "extrapolation_unreliable": {
            "type": "boolean"
          },
          "forecast_data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ForecastPoint"
            }
          },
          "contradictions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ContradictionRecord"
            }
          }
        }
      },
      "PipelineTrace": {
        "type": "object",
        "description": "Full pipeline trace returned by GET /api/pipeline/:id and /api/pipeline/:id/trace. Mirrors src/tracing/collector.ts.",
        "required": [
          "pipeline_id",
          "environment",
          "ai_provider_used",
          "started_at",
          "workplan",
          "task_plan",
          "reasoning_steps",
          "tool_calls",
          "action_execution",
          "recovery_steps",
          "events"
        ],
        "properties": {
          "pipeline_id": {
            "type": "string",
            "pattern": "^PIPE-[A-Z0-9]{8}$"
          },
          "environment": {
            "type": "string",
            "example": "development"
          },
          "ai_provider_used": {
            "type": "string",
            "example": "gemini-free"
          },
          "started_at": {
            "type": "string",
            "format": "date-time"
          },
          "completed_at": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "workplan": {
            "type": "string"
          },
          "task_plan": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "reasoning_steps": {
            "type": "array",
            "items": {
              "type": "object",
              "required": [
                "step",
                "agent",
                "reasoning",
                "decision",
                "confidence"
              ],
              "properties": {
                "step": {
                  "type": "integer",
                  "minimum": 1
                },
                "agent": {
                  "type": "string"
                },
                "reasoning": {
                  "type": "string"
                },
                "decision": {
                  "type": "string"
                },
                "confidence": {
                  "type": "number",
                  "minimum": 0,
                  "maximum": 1
                }
              }
            }
          },
          "tool_calls": {
            "type": "array",
            "items": {
              "type": "object",
              "required": [
                "tool",
                "provider",
                "input_summary",
                "output_summary",
                "latency_ms"
              ],
              "properties": {
                "tool": {
                  "type": "string"
                },
                "provider": {
                  "type": "string"
                },
                "input_summary": {
                  "type": "string"
                },
                "output_summary": {
                  "type": "string"
                },
                "latency_ms": {
                  "type": "number",
                  "minimum": 0
                }
              }
            }
          },
          "action_execution": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TraceEvent"
            }
          },
          "recovery_steps": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "events": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TraceEvent"
            }
          }
        }
      },
      "PipelineRunRequest": {
        "type": "object",
        "required": [
          "sources"
        ],
        "properties": {
          "sources": {
            "type": "array",
            "minItems": 1,
            "items": {
              "$ref": "#/components/schemas/MultiSourceIngestionSource"
            }
          },
          "constraints": {
            "type": "object",
            "additionalProperties": true,
            "example": {
              "max_actions": 10,
              "forbidden_action_types": [
                "IRREVERSIBLE_DELETE"
              ]
            }
          }
        }
      },
      "PipelineRunResponse": {
        "type": "object",
        "required": [
          "pipeline_id",
          "status"
        ],
        "properties": {
          "pipeline_id": {
            "type": "string",
            "pattern": "^PIPE-[A-Z0-9]{8}$",
            "example": "PIPE-7K2A9F31"
          },
          "status": {
            "type": "string",
            "example": "received"
          },
          "message": {
            "type": "string"
          },
          "sources_received": {
            "type": "integer",
            "minimum": 0
          },
          "constraints_received": {
            "type": "object",
            "additionalProperties": true
          }
        }
      },
...
    }
  }
}